More nonfunctional stuff
Added view invoices screen and the required default in the database. I
am still unsure if I should store notes over multiple variables for each
date, or for the whole.  Most of the screens exist, with a few done. The
only screen to be completely added is the add a client screen